# nightclub-app/spec/current/openapi.yaml
openapi: 3.1.1
info:
  title: Nightclub v4 R1 Core API Design
  version: 4.0.0
  description: 開発用コア契約。未実装。ユーザー確定：名前検索中心、入口も通常承認、入口共用端末。R1全機能の詳細化状況はapi/coverage.mdを参照。実サービスURL・認証基盤・PSPは未確定。
servers:
- url: https://api.example.invalid/v1
  description: 非実在の例示URL。実接続先ではない。
security: &id001
- PersonalSession: []
- DeviceSession: []
  OperatorSession: []
paths:
  /auth/line/start:
    get:
      operationId: startLineLogin
      tags:
      - Core
      summary: 個人LINEログイン開始
      description: 'state・nonce・PKCE S256を生成。callbackは選定認証基盤の認証設計で別管理。


        権限：public。公開経路でもレート制限・適用範囲・冪等性をサーバーで検証。'
      security: []
      x-permission: public
      x-requirement-ids:
      - V3-ID-02
      responses:
        '200':
          description: 成功（外部処理待ちは状態で明示）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginStart'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /me:
    get:
      operationId: getMe
      tags:
      - Core
      summary: 現在の主体と権限
      description: '


        権限：session.read。入口モードでは有効な端末・担当者・当日役割を検証。主体IDをbodyで偽装できない。'
      security: *id001
      x-permission: session.read
      x-requirement-ids:
      - V3-ID-01
      - V4-KSK-03
      responses:
        '200':
          description: 成功（外部処理待ちは状態で明示）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActorContext'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
      parameters:
      - $ref: '#/components/parameters/OperatorContext'
  /auth/logout:
    post:
      operationId: logout
      tags:
      - Core
      summary: セッション失効
      description: '


        権限：session.end。入口モードでは有効な端末・担当者・当日役割を検証。主体IDをbodyで偽装できない。'
      security: *id001
      x-permission: session.end
      x-requirement-ids:
      - V3-ID-06
      responses:
        '200':
          description: 成功（外部処理待ちは状態で明示）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ack'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      - $ref: '#/components/parameters/OperatorContext'
...
  /stores/{storeId}/memberships:
    get:
      operationId: listMemberships
      tags:
      - Admin
      summary: 店舗所属一覧
      description: '


        権限：membership.manage。入口モードでは有効な端末・担当者・当日役割を検証。主体IDをbodyで偽装できない。'
      security: *id001
      x-permission: membership.manage
      x-requirement-ids: []
      responses:
        '200':
          description: 成功（外部処理待ちは状態で明示）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipSummaryPage'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
      parameters:
      - name: storeId
        in: path
        required: true
        schema: *id003
      - $ref: '#/components/parameters/OperatorContext'
  /stores/{storeId}/invitations:
    post:
      operationId: inviteMember
      tags:
      - Admin
      summary: 期限付き個人招待
      description: '


        権限：membership.manage。入口モードでは有効な端末・担当者・当日役割を検証。主体IDをbodyで偽装できない。'
      security: *id001
      x-permission: membership.manage
      x-requirement-ids:
      - V3-ID-03
      responses:
        '201':
          description: 成功（外部処理待ちは状態で明示）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResult'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
      parameters:
      - name: storeId
        in: path
        required: true
        schema: *id003
      - $ref: '#/components/parameters/CsrfToken'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/OperatorContext'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteCreate'
  /stores/{storeId}/memberships/{membershipId}/suspend:
    post:
      operationId: suspendMember
      tags:
      - Admin
      summary: 所属と有効セッションを停止
      description: '


        権限：membership.manage。入口モードでは有効な端末・担当者・当日役割を検証。主体IDをbodyで偽装できない。'
      security: *id001
      x-permission: membership.manage
      x-requirement-ids:
      - V3-ID-06
      responses:
        '200':
          description: 成功（外部処理待ちは状態で明示）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResult'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
      parameters:
      - name: storeId
        in: path
        required: true
        schema: *id003
      - name: membershipId
        in: path
        required: true
        schema: *id003
      - $ref: '#/components/parameters/CsrfToken'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/OperatorContext'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonCommand'
  /stores/{storeId}/roles/{roleId}:
    put:
      operationId: updateRole
      tags:
      - Admin
      summary: 店舗ロール権限を更新
      description: '


        権限：role.manage。入口モードでは有効な端末・担当者・当日役割を検証。主体IDをbodyで偽装できない。'
      security: *id001
      x-permission: role.manage
      x-requirement-ids:
      - V3-ID-09
      responses:
        '200':
          description: 成功（外部処理待ちは状態で明示）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResult'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
      parameters:
      - name: storeId
        in: path
        required: true
        schema: *id003
      - name: roleId
        in: path
        required: true
        schema: *id003
      - $ref: '#/components/parameters/CsrfToken'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/OperatorContext'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleUpdate'
...
  /stores/{storeId}/memberships/{membershipId}/operator-pin:
    put:
      operationId: resetOperatorPin
      tags:
      - Identity
      summary: 共用端末用の個人PINを設定・再発行
      description: '追加認証した管理者のみ。PINはwriteOnly・ログ禁止。既存operatorセッションを失効させる。実際の初回配布方法は認証設計審査で確定。


        権限：credential.manage。入口モードでは有効な端末・担当者・当日役割を検証。主体IDをbodyで偽装できない。'
      security: *id001
      x-permission: credential.manage
      x-requirement-ids:
      - V4-KSK-02
      - V4-KSK-05
      responses:
        '200':
          description: 成功（外部処理待ちは状態で明示）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ack'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
      parameters:
      - name: storeId
        in: path
        required: true
        schema: *id003
      - name: membershipId
        in: path
        required: true
        schema: *id003
      - $ref: '#/components/parameters/CsrfToken'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/OperatorContext'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PinReset'
...
components:
  securitySchemes:
    PersonalSession:
      type: apiKey
      in: cookie
      name: __Host-nc_personal
    DeviceSession:
      type: apiKey
      in: cookie
      name: __Host-nc_device
    OperatorSession:
      type: apiKey
      in: cookie
      name: __Host-nc_operator
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 160
      description: 同じ業務操作は同じキー。異なるpayloadへの流用は409。金銭は保存済み参照も照合。
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: *id002
      description: cookie認証の状態変更に必要。Origin/Fetch Metadataも検証。
    OperatorContext:
      name: X-Operator-Context
      in: header
      required: false
      schema: *id003
      description: 入口モードでは必須。画面描画時のoperator_session_id。認証手段ではない。cookie主体とサーバーactive sessionに一致しなければ409 OPERATOR_CHANGED。個人モードでは不要。
  schemas:
    Problem:
      type: object
      properties:
        type: *id002
        title: *id002
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail: *id002
        code: *id002
        trace_id: *id003
        current_version:
          type:
          - integer
          - 'null'
        retryable:
          type: boolean
        field_errors:
          type: array
          items:
            type: object
            properties:
              path: *id002
              code: *id002
              message: *id002
            required:
            - path
            - code
            - message
            additionalProperties: false
      required:
      - type
      - title
      - status
      - code
      - trace_id
      - retryable
      additionalProperties: false
    Ack:
      type: object
      properties:
        accepted:
          type: boolean
        trace_id: *id003
      required:
      - accepted
      - trace_id
      additionalProperties: false
    CommandResult:
      type: object
      properties:
        resource_id: *id003
        version: &id005
          type: integer
          minimum: 1
        status: *id002
        trace_id: *id003
        operation_key: *id002
      required:
      - resource_id
      - version
      - status
      - trace_id
      additionalProperties: false
...
    ReasonCommand:
      type: object
      properties:
        expected_version: *id005
        reason:
          type: string
          minLength: 1
          maxLength: 500
      required:
      - expected_version
      - reason
      additionalProperties: false
    ActorContext:
      type: object
      properties:
        user_id: *id003
        membership_id: *id003
        store_id: *id003
        display_name: *id002
        permissions:
          type: array
          items: *id002
        mode:
          type: string
          enum:
          - PERSONAL
          - ENTRANCE
        operator_session_id: &id006
          type:
          - string
          - 'null'
          format: uuid
        device_id: *id006
        event_id: *id006
      required:
      - user_id
      - membership_id
      - store_id
      - display_name
      - permissions
      - mode
      - operator_session_id
      - device_id
      - event_id
      additionalProperties: false
    LoginStart:
      type: object
      properties:
        authorization_url:
          type: string
          format: uri
        expires_at: &id007
          type: string
          format: date-time
      required:
      - authorization_url
      - expires_at
      additionalProperties: false
...
    MembershipSummary:
      type: object
      properties:
        id: *id003
        version: *id005
        display_name: *id002
        status: *id002
        role_keys:
          type: array
          items: *id002
      required:
      - id
      - version
      - display_name
      - status
      - role_keys
      additionalProperties: false
...
    InviteCreate:
      type: object
      properties:
        target:
          type: string
          minLength: 1
          maxLength: 200
        role_id: *id003
        expires_at: *id007
      required:
      - target
      - role_id
      - expires_at
      additionalProperties: false
    RoleUpdate:
      type: object
      properties:
        expected_version: *id005
        permission_keys:
          type: array
          items: *id002
          uniqueItems: true
      required:
      - expected_version
      - permission_keys
      additionalProperties: false
...
    MembershipSummaryPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MembershipSummary'
        next_cursor:
          type:
          - string
          - 'null'
      required:
      - items
      - next_cursor
      additionalProperties: false
...
    PinReset:
      type: object
      properties:
        expected_membership_version: *id005
        new_pin:
          type: string
          pattern: ^[0-9]{6,12}$
          writeOnly: true
        reason:
          type: string
          minLength: 1
          maxLength: 500
      required:
      - expected_membership_version
      - new_pin
      - reason
      additionalProperties: false
...
  responses:
    Problem:
      description: 構造化エラー。detailsで個人情報・内部SQLを漏らさない。
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'